feat: liveness report, removing idle connections
implement the followings:
- update_peer_state
- remove_idle_peers

test:
system_removes_node_when_heartbeat_is_not_received_for_certain_time

backlog - use of hop count

+++ src/services/config/init.rs
@@ -9,7 +9,7 @@ pub(crate) struct Environment {
     pub(crate) port: u16,
     pub(crate) host: String,
     pub(crate) hf_mills: u64,
-    pub(crate) ttl_mills: u64,
+    pub(crate) ttl_mills: u128,
 }
 
 impl Environment {
